+++ util.go
@@ -1,0 +1,29 @@
+package main
+
+import (
+	"errors"
+	"fmt"
+	"net"
+	"os"
+)
+
+func getDefaultInterface() (string, error) {
+	interfaces, err := net.Interfaces()
+	if err != nil {
+		return "", err
+	}
+	for _, iface := range interfaces {
+		if iface.Flags&net.FlagLoopback == 0 && iface.Flags&net.FlagUp != 0 {
+			return iface.Name, nil
+		}
+	}
+	return "", errors.New("no active non-loopback interface found")
+}
+
+func getHostname() (string, error) {
+	hostname, err := os.Hostname()
+	if err != nil {
+		return "", fmt.Errorf("failed to get hostname: %s", err)
+	}
+	return hostname, nil
+}
